@@ -6,7 +6,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0-alpha.1+1
 
 environment:
-  sdk: ">=3.0.3 <4.0.0"
+  sdk: ">=3.7.2 <4.0.0"
 
 dependencies:
   flutter:
@@ -15,68 +15,52 @@ dependencies:
     sdk: flutter
 
   # Architecture
-  flutter_bloc: ^8.1.3
-  go_router: ^12.1.1
-  dio: ^5.3.4
-  get_it: ^7.6.4
-  flutter_secure_storage: ^9.0.0
-  shared_preferences: ^2.2.2
-  easy_localization: ^3.0.3
-  equatable: ^2.0.5
-  json_annotation: ^4.8.1
-  connectivity_plus: ^4.0.2
+  flutter_bloc: ^9.1.1
+  go_router: ^17.0.1
+  dio: ^5.9.0
+  get_it: ^9.2.0
+  flutter_secure_storage: ^10.0.0
+  shared_preferences: ^2.5.4
+  easy_localization: ^3.0.8
+  equatable: ^2.0.7
+  json_annotation: ^4.9.0
+  connectivity_plus: ^7.0.0
 
   # File Manager
-  path_provider: ^2.1.1
-  image_picker: ^1.0.4
-  file_picker: ^6.1.1
-  flutter_image_compress: ^2.1.0
-  path: ^1.8.3
+  path_provider: ^2.1.5
+  image_picker: ^1.2.1
+  file_picker: ^10.3.8
+  flutter_image_compress: ^2.4.0
+  path: ^1.9.1
 
   # Widget
-  modal_bottom_sheet: ^3.0.0-pre
+  modal_bottom_sheet: ^3.0.0
   infinite_scroll_pagination: ^4.0.0
   shimmer: ^3.0.0
-  cached_network_image: ^3.3.0
-  cupertino_icons: ^1.0.6
-  flutter_svg: ^2.0.9
+  cached_network_image: ^3.4.1
+  cupertino_icons: ^1.0.8
+  flutter_svg: ^2.2.3
 
   # Others
-  package_info_plus: ^5.0.1
-  device_info_plus: ^9.0.2
-  url_launcher: ^6.2.1
-  permission_handler: ^11.1.0
-  intl: ^0.18.1
-  open_filex: ^4.3.4
-  device_preview: ^1.1.0
+  collection: ^1.19.1
+  package_info_plus: ^9.0.0
+  device_info_plus: ^12.3.0
+  url_launcher: ^6.3.2
+  permission_handler: ^12.0.1
+  intl: ^0.20.2
+  open_filex: ^4.7.0
+  device_preview: ^1.3.1
+  sliver_tools: ^0.2.12
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^3.0.1
-  flutter_native_splash: ^2.3.6
-
-#  change_app_package_name: ^1.1.0
-#  Usage :
-#  dart run change_app_package_name:main com.new.package.name
-
-#  rename: ^3.0.1
-#  Usage :
-#  rename setAppName --targets ios,android --value "YourAppName"
-#  rename setBundleId --targets android --value "com.example.bundleId"
-
-#  flutter_launcher_icons: ^0.13.1
-#  Usage :
-#  flutter pub run flutter_launcher_icons
-#
-#flutter_icons:
-#  image_path: "assets/images/logo/app_logo.png"
-#  min_sdk_android: 21
-#  ios: true
-#  android: true
-#  remove_alpha_ios: true (Optional)
-#  background_color_ios: "#ffffff" (Optional)
+  flutter_lints: ^5.0.0
+  flutter_native_splash: ^2.4.7
+  change_app_package_name: ^1.5.0
+#  change_app_package_name: ^1.4.0
+#  flutter_launcher_icons: ^0.14.1
 
 flutter:
   uses-material-design: true
@@ -99,3 +83,15 @@ flutter:
           weight: 400
         - asset: fonts/Poppins-Light.ttf
           weight: 300
+
+#
+# Usage :
+# flutter pub run flutter_launcher_icons
+#
+#flutter_icons:
+#  image_path: "assets/images/logo/app_logo.png"
+#  min_sdk_android: 21
+#  ios: true
+#  android: true
+#  remove_alpha_ios: true (Optional)
+#  background_color_ios: "#ffffff" (Optional)
