@@ -24,7 +24,6 @@ class _MainNavViewState extends State<MainNavView> {
   Widget build(BuildContext context) {
     return Scaffold(
       body: ColoredStatusBar.primary(
-        brightness: Brightness.dark,
         child: SafeArea(
           child: PageView(
             controller: pageController,
@@ -48,7 +47,7 @@ class _MainNavViewState extends State<MainNavView> {
         borderColor: Colors.transparent,
         boxShadow: [
           BoxShadow(
-            color: Colors.grey.withOpacity(0.2),
+            color: Colors.grey.withValues(alpha: 0.2),
             spreadRadius: 5,
             blurRadius: 3,
             offset: const Offset(0, 3), // changes position of shadow
