@@ -7,7 +7,11 @@ import 'package:package_info_plus/package_info_plus.dart';
 class AppInfo {
   static late PackageInfo packageInfo;
 
-  static setInfo(PackageInfo info) {
+  static void setInfo(PackageInfo info) {
     packageInfo = info;
   }
+
+  static String appVersion = AppInfo.packageInfo.version;
+  static String buildNumber = AppInfo.packageInfo.buildNumber;
+  static String packageName = AppInfo.packageInfo.packageName;
 }
