@@ -6,18 +6,22 @@
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:skybase/config/base/connectivity_mixin.dart';
-import 'package:skybase/core/database/storage/storage_manager.dart';
+import 'package:skybase/config/base/request_param.dart';
+import 'package:skybase/core/mixin/cache_mixin.dart';
+import 'package:skybase/core/mixin/connectivity_mixin.dart';
 
-class BaseCubit<S, T> extends Cubit<S> with ConnectivityMixin {
+class BaseCubit<S, T> extends Cubit<S> with ConnectivityMixin, CacheMixin {
   BaseCubit(super.initialState);
 
-  StorageManager storage = StorageManager.instance;
-
   CancelToken cancelToken = CancelToken();
+  late RequestParams requestParams;
   String? errorMessage;
   bool isError = false;
   bool isLoading = false;
+
+  String get cachedKey => '';
+
+  String get cachedId => '';
 
   void loadData(Function() onLoad) {
     onLoad();
@@ -25,6 +29,11 @@ class BaseCubit<S, T> extends Cubit<S> with ConnectivityMixin {
 
   @mustCallSuper
   void onInit([dynamic args]) {
+    requestParams = RequestParams(
+      cancelToken: cancelToken,
+      cachedKey: cachedKey,
+      cachedId: cachedId,
+    );
     listenConnectivity(() {
       if (isError && !isLoading) onRefresh();
     });
@@ -32,13 +41,13 @@ class BaseCubit<S, T> extends Cubit<S> with ConnectivityMixin {
 
   void onRefresh([BuildContext? context]) {}
 
-  Future<void> deleteCached(String cacheKey, {String? cacheId}) async {
-    if (cacheId != null) {
-      await storage.delete('$cacheKey/$cacheId');
-    } else {
-      await storage.delete(cacheKey.toString());
-    }
-  }
+  // Future<void> deleteCached(String cacheKey, {String? cacheId}) async {
+  //   if (cacheId != null) {
+  //     await storage.delete('$cacheKey/$cacheId');
+  //   } else {
+  //     await storage.delete(cacheKey.toString());
+  //   }
+  // }
 
   void emitLoading(S state) {
     isLoading = true;
