@@ -22,7 +22,7 @@ sealed class NetworkExceptionData with ApiMessage {
   @override
   String toString() {
     String result = '';
-    if (response?.statusCode == 400 || response?.statusCode == 401) {
+    if (response?.statusCode == 400) {
       ApiResponse res = ApiResponse.fromJson(response?.data);
       result = convertMessage(res.error ?? res.message);
     } else {
@@ -36,17 +36,17 @@ mixin NetworkException implements Exception {
   NetworkExceptionData handleResponse(Response response) {
     int statusCode = response.statusCode!;
     return switch (statusCode) {
-      400 || 403 || 422  => BadRequestException(response: response),
-      401 => UnauthorisedException(response: response),
-      404 => NotFoundException(response: response),
-      409 => FetchDataException(response: response),
+      400 || 403 || 422 => BadRequestException(response: response),
+      401 => UnauthorisedException(),
+      404 => NotFoundException(),
+      409 => FetchDataException(),
       408 => SendTimeOutException(),
-      413 => RequestEntityTooLargeException(response: response),
+      413 => RequestEntityTooLargeException(),
       500 || 503 => InternalServerErrorException(),
       _ => FetchDataException(
-        message: 'Received invalid status code: $statusCode',
-        response: response,
-      ),
+          message: 'Received invalid status code: $statusCode',
+          response: response,
+        ),
     };
   }
 
@@ -112,8 +112,7 @@ final class RequestEntityTooLargeException extends NetworkExceptionData {
 
 final class FetchDataException extends NetworkExceptionData {
   FetchDataException({String? message, super.response})
-      : super(
-      message: message ?? 'txt_error_during_communication'.tr());
+      : super(message: message ?? 'txt_error_during_communication'.tr());
 }
 
 final class NotFoundException extends NetworkExceptionData {
@@ -124,8 +123,8 @@ final class NotFoundException extends NetworkExceptionData {
 final class BadRequestException extends NetworkExceptionData {
   BadRequestException({super.response})
       : super(
-      prefix: 'txt_bad_request'.tr(),
-      message: '${'txt_message'.tr()}: ${response?.statusMessage}');
+            prefix: 'txt_bad_request'.tr(),
+            message: '${'txt_message'.tr()}: ${response?.statusMessage}');
 }
 
 final class BadCertificateException extends NetworkExceptionData {
