@@ -1,4 +1,4 @@
-import 'package:dio/dio.dart';
+import 'package:skybase/config/base/request_param.dart';
 import 'package:skybase/data/models/repo/repo.dart';
 import 'package:skybase/data/models/user/user.dart';
 
@@ -10,16 +10,17 @@ abstract interface class AuthRepository {
 
   Future<User> login({
     required String phoneNumber,
+    required String email,
     required String password,
   });
 
   Future<User> getProfile({
-    required CancelToken cancelToken,
+    required RequestParams requestParams,
     required String username,
   });
 
   Future<List<Repo>> getProfileRepository({
-    required CancelToken cancelToken,
+    required RequestParams requestParams,
     required String username,
   });
 }
