@@ -1,14 +1,12 @@
 import 'dart:developer';
 
-import 'package:dio/dio.dart';
-import 'package:skybase/config/base/cache_mixin.dart';
+import 'package:skybase/config/base/base_repository.dart';
+import 'package:skybase/config/base/request_param.dart';
 import 'package:skybase/data/models/sample_feature/sample_feature.dart';
 import 'package:skybase/data/repositories/sample_feature/sample_feature_repository.dart';
-import 'package:skybase/data/sources/local/cached_key.dart';
 import 'package:skybase/data/sources/server/sample_feature/sample_feature_sources.dart';
 
-class SampleFeatureRepositoryImpl
-    with CacheMixin
+class SampleFeatureRepositoryImpl extends BaseRepository
     implements SampleFeatureRepository {
   String tag = 'SampleFeatureRepository::->';
 
@@ -18,24 +16,31 @@ class SampleFeatureRepositoryImpl
 
   @override
   Future<List<SampleFeature>> getUsers({
-    required CancelToken cancelToken,
+    required RequestParams requestParams,
     required int page,
     required int perPage,
+    String? username,
   }) async {
     try {
       // Using cached
-      return await getCacheList(
-        cachedKey: CachedKey.SAMPLE_FEATURE_LIST,
-        page: page,
+      return await loadCachedList(
+        cachedKey: requestParams.cachedKey.toString(),
+        loadWhen: page == 1 && username == null,
         onLoad: () async => await apiService.getUsers(
-          cancelToken: cancelToken,
+          cancelToken: requestParams.cancelToken,
           page: page,
           perPage: perPage,
+          username: username,
         ),
       );
 
       // Without cache
-      // return await apiService.getUsers(page: page, perPage: perPage);
+      // return await apiService.getUsers(
+      //   cancelToken: requestParams.cancelToken,
+      //   page: page,
+      //   perPage: perPage,
+      //   username: username,
+      // );
     } catch (e, stack) {
       log('$tag error = $e, $stack');
       rethrow;
@@ -44,29 +49,29 @@ class SampleFeatureRepositoryImpl
 
   @override
   Future<SampleFeature> getDetailUser({
-    required CancelToken cancelToken,
+    required RequestParams requestParams,
     required int id,
     required String username,
   }) async {
     try {
       // Using cache
-      return await getCache(
-        cachedKey: CachedKey.SAMPLE_FEATURE_DETAIL,
-        cachedId: id.toString(),
+      return await loadCached(
+        cachedKey: requestParams.cachedKey.toString(),
+        cachedId: requestParams.cachedId,
         onLoad: () async => await apiService
-            .getDetailUser(cancelToken: cancelToken, username: username)
+            .getDetailUser(cancelToken: requestParams.cancelToken, username: username)
             .then(
           (res) async {
             res.followersList = await apiService.getFollowers(
-              cancelToken: cancelToken,
+              cancelToken: requestParams.cancelToken,
               username: username,
             );
             res.followingList = await apiService.getFollowings(
-              cancelToken: cancelToken,
+              cancelToken: requestParams.cancelToken,
               username: username,
             );
             res.repositoryList = await apiService.getRepos(
-              cancelToken: cancelToken,
+              cancelToken: requestParams.cancelToken,
               username: username,
             );
             return res;
@@ -75,10 +80,22 @@ class SampleFeatureRepositoryImpl
       );
 
       // Without Cache
-      // final SampleFeature res = await apiService.getDetailUser(username: username);
-      // res.followersList = await apiService.getFollowers(username: username);
-      // res.followingList = await apiService.getFollowings(username: username);
-      // res.repositoryList = await apiService.getRepos(username: username);
+      // final SampleFeature res = await apiService.getDetailUser(
+      //   cancelToken: requestParams.cancelToken,
+      //   username: username,
+      // );
+      // res.followersList = await apiService.getFollowers(
+      //   cancelToken: requestParams.cancelToken,
+      //   username: username,
+      // );
+      // res.followingList = await apiService.getFollowings(
+      //   cancelToken: requestParams.cancelToken,
+      //   username: username,
+      // );
+      // res.repositoryList = await apiService.getRepos(
+      //   cancelToken: requestParams.cancelToken,
+      //   username: username,
+      // );
       // return res;
     } catch (e, stack) {
       log('$tag Failed get data $e, $stack');
