@@ -1,9 +1,10 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 import 'package:skybase/service_locator.dart';
-import 'package:skybase/ui/views/profile/cubit/profile_cubit.dart';
-import 'package:skybase/ui/views/profile/component/repository/cubit/profile_repository_cubit.dart';
 import 'package:skybase/ui/views/profile/profile_view.dart';
+
+import 'component/repository/cubit/profile_repository_cubit.dart';
+import 'cubit/profile_cubit.dart';
 
 final profileRoute = [
   GoRoute(
@@ -17,6 +18,6 @@ final profileRoute = [
         ],
         child: const ProfileView(),
       );
-    },
+    }
   ),
 ];
