@@ -2,13 +2,14 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import 'package:skybase/app_configuration.dart';
+import 'package:skybase/config/app/app_info.dart';
 import 'package:skybase/config/themes/app_style.dart';
-import 'package:skybase/config/themes/theme_manager/theme_manager.dart';
-import 'package:skybase/ui/views/settings/cubit/setting_cubit.dart';
+import 'package:skybase/config/themes/theme_manager.dart';
 import 'package:skybase/ui/widgets/colored_status_bar.dart';
 import 'package:skybase/ui/widgets/sky_appbar.dart';
 import 'package:skybase/ui/widgets/sky_button.dart';
+
+import 'cubit/setting_cubit.dart';
 
 class SettingView extends StatefulWidget {
   static const String route = '/setting';
@@ -29,7 +30,6 @@ class _SettingViewState extends State<SettingView> {
   @override
   Widget build(BuildContext context) {
     return ColoredStatusBar.primary(
-      brightness: Brightness.dark,
       child: Scaffold(
         appBar: SkyAppBar.secondary(title: 'txt_setting'.tr()),
         bottomNavigationBar: Padding(
@@ -43,7 +43,7 @@ class _SettingViewState extends State<SettingView> {
             mainAxisSize: MainAxisSize.min,
             children: [
               Text(
-                '${'txt_version'.tr()} ${AppConfiguration.appVersion}',
+                '${'txt_version'.tr()} ${AppInfo.appVersion}',
                 style: AppStyle.body2.copyWith(color: Colors.grey),
               ),
               const SizedBox(height: 12),
@@ -71,14 +71,14 @@ class _SettingViewState extends State<SettingView> {
                           mainAxisAlignment: MainAxisAlignment.end,
                           children: [
                             const Text('ENG'),
-                            Radio(
+                            Radio<String>(
                               value: 'en',
                               groupValue: state.languageCode,
                               onChanged: (value) {
                                 context.read<SettingCubit>().onUpdateLocale(
-                                      context: context,
-                                      languageCode: value.toString(),
-                                    );
+                                  context: context,
+                                  languageCode: value.toString(),
+                                );
                               },
                             ),
                             const Text('ID'),
@@ -87,16 +87,16 @@ class _SettingViewState extends State<SettingView> {
                               groupValue: state.languageCode,
                               onChanged: (value) async {
                                 context.read<SettingCubit>().onUpdateLocale(
-                                      context: context,
-                                      languageCode: value.toString(),
-                                    );
+                                  context: context,
+                                  languageCode: value.toString(),
+                                );
                               },
                             ),
                           ],
                         );
                       },
                     ),
-                  )
+                  ),
                 ],
               ),
               const Divider(color: Colors.grey),
@@ -105,12 +105,14 @@ class _SettingViewState extends State<SettingView> {
                 children: [
                   Text('txt_dark_mode'.tr()),
                   BlocBuilder<ThemeManager, ThemeState>(
-                    builder: (context, state) => Switch(
-                      value: state is IsDarkMode,
-                      onChanged: (value) {
-                        context.read<ThemeManager>().changeTheme();
-                      },
-                    ),
+                    builder: (context, state) {
+                      return Switch(
+                        value: state is IsDarkMode,
+                        onChanged: (value) {
+                          context.read<ThemeManager>().changeTheme();
+                        },
+                      );
+                    },
                   ),
                 ],
               ),
