@@ -2,10 +2,9 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:skybase/core/helper/converter_helper.dart';
-import 'package:skybase/core/extension/int_extension.dart';
+import 'package:skybase/core/extension/num_extension.dart';
 import 'package:skybase/core/helper/input_formater.dart';
 import 'package:skybase/core/helper/snackbar_helper.dart';
-import 'package:skybase/core/helper/validator_helper.dart';
 import 'package:skybase/config/themes/app_style.dart';
 import 'package:skybase/ui/widgets/keyboard_dismissible.dart';
 import 'package:skybase/ui/widgets/sky_appbar.dart';
@@ -45,7 +44,7 @@ class _OtherUtilsViewState extends State<OtherUtilsView> {
                 onChanged: (value) => (value.isEmpty)
                     ? currencyCtr.text = 0.currencyFormat(symbol: 'Rp')
                     : value,
-                validator: (value) => ValidatorHelper.required('$value'),
+                isRequired: true,
                 inputFormatters: CustomInputFormatters.idrCurrency,
                 onFieldSubmitted: (value) =>
                     SnackBarHelper.normal(context: context, message: value),
