@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:skybase/config/themes/theme_manager/theme_manager.dart';
+import 'package:skybase/config/themes/theme_manager.dart';
 import 'package:skybase/ui/widgets/sky_appbar.dart';
 
 class ThemeComponentUtilsView extends StatefulWidget {
