@@ -17,6 +17,8 @@ class PaginationSliverGrid<T> extends StatelessWidget {
     required this.pagingController,
     required this.itemBuilder,
     required this.gridDelegate,
+    required this.onRetry,
+    this.onRefresh,
     this.padding,
     this.showNewPageErrorIndicatorAsGridChild = false,
     this.showNewPageProgressIndicatorAsGridChild = false,
@@ -27,7 +29,7 @@ class PaginationSliverGrid<T> extends StatelessWidget {
     this.loadingView,
     this.emptyView,
     this.errorView,
-    this.errorLoadView,
+    this.errorLoadMoreView,
     this.maxItemView,
     this.emptyImage,
     this.emptyTitle,
@@ -47,12 +49,12 @@ class PaginationSliverGrid<T> extends StatelessWidget {
     this.emptyImageWidget,
     this.emptyTitleStyle,
     this.emptySubtitleStyle,
-    this.onRefresh,
   });
 
   final PagingController<int, T> pagingController;
   final ItemWidgetBuilder<T> itemBuilder;
   final SliverGridDelegate gridDelegate;
+  final VoidCallback onRetry;
   final bool showNewPageErrorIndicatorAsGridChild;
   final bool showNewPageProgressIndicatorAsGridChild;
   final bool showNoMoreItemsIndicatorAsGridChild;
@@ -64,7 +66,7 @@ class PaginationSliverGrid<T> extends StatelessWidget {
   final Widget? emptyView;
   final Widget? maxItemView;
   final Widget? errorView;
-  final Widget? errorLoadView;
+  final Widget? errorLoadMoreView;
   final VoidCallback? onRefresh;
   final Widget? emptyImageWidget;
   final String? emptyImage;
@@ -108,6 +110,7 @@ class PaginationSliverGrid<T> extends StatelessWidget {
         shrinkWrapFirstPageIndicators: shrinkWrapFirstPageIndicators,
         builderDelegate: PaginationDelegate<T>(
           pagingController: pagingController,
+          onRetry: onRetry,
           loadingView: loadingView,
           emptyView: emptyView,
           emptyRetryEnabled: emptyRetryEnabled,
@@ -131,7 +134,7 @@ class PaginationSliverGrid<T> extends StatelessWidget {
           imageHeight: imageHeight,
           imageWidth: imageWidth,
           itemBuilder: itemBuilder,
-          errorLoadView: errorLoadView,
+          errorLoadMoreView: errorLoadMoreView,
           errorView: errorView,
         ),
       ),
